+++ scripts/exwafs_grib2_0p25_blending.sh
@@ -18,35 +18,40 @@ set -x
 
 cd "${DATA}" || err_exit "FATAL ERROR: Could not 'cd ${DATA}'; ABORT!"
 
-fhours=${fhours:-"006 007 008 009 010 011 012 013 014 015 016 017 018 019 020 021 022 023 024 027 030 033 036 039 042 045 048"}
-np=`echo $fhours | wc -w`
-MPIRUN="mpiexec -np $np -cpu-bind verbose,core cfp"
+fhours=($(seq -s ' ' -f "%03g" 6 1 24; seq -s ' ' -f "%03g" 27 3 48))
+np=${#fhours[@]}
+MPIRUN="mpiexec -np ${np} -cpu-bind verbose,core cfp"
 
 rm -f wafsgrib2_0p25.cmdfile
-ic=0
-for fhr in $fhours ; do
-  if [[ $(echo $MPIRUN | cut -d " " -f1) = 'srun' ]] ; then
-    echo "$ic ${USHwafs}/wafs_grib2_0p25_blending.sh $fhr > $DATA/${fhr}.log 2>&1" >> wafsgrib2_0p25.cmdfile
-  else
-    echo "${USHwafs}/wafs_grib2_0p25_blending.sh $fhr > $DATA/${fhr}.log 2>&1">> wafsgrib2_0p25.cmdfile
+for fhr in ${fhours[@]}; do
+    echo "${USHwafs}/wafs_grib2_0p25_blending.sh $fhr > $DATA/${fhr}.log 2>&1">> wafsgrib2_0p25_blending.cmdfile
     export MP_PGMMODEL=mpmd
-  fi
-  ic=$(expr $ic + 1)
 done
-$MPIRUN wafsgrib2_0p25.cmdfile
+$MPIRUN wafsgrib2_0p25_blending.cmdfile
+
+export err=$? ; err_chk
+
+for fhr in ${fhours[@]}; do
+    echo "=================== log file of fhr=$fhr ==================="
+    cat "${DATA}/${fhr}.log"
+done
+echo "===================== end of log files ====================="
 
 missing_uk_files="$(find $DATA -name 'missing_uk_files*')"
+missing_us_files="$(find $DATA -name 'missing_us_file*')"
+no_blending_files="$(find $DATA -name 'no_blending_files*')"
 
-if [[ ! -z "$missing_uk_files" ]] ; then
-
-    echo "WARNING: Missing UK data for WAFS GRIB2 0P25 blending. Send alert message to AWC ......"
+if [[ ! -z "$missing_uk_files" ]] || [[ ! -z "$missing_us_files" ]] || [[ ! -z "$no_blending_files" ]] ; then
+    echo "WARNING: No WAFS GRIB2 0P25 blending. Send alert message to AWC ......"
     make_NTC_file.pl NOXX10 KKCI "${PDY}${cyc}" NONE "${FIXwafs}/wafs/wafs_blending_0p25_admin_msg" "${COMOUTwmo}/wifs_0p25_admin_msg"
     make_NTC_file.pl NOXX10 KWBC "${PDY}${cyc}" NONE "${FIXwafs}/wafs/wafs_blending_0p25_admin_msg" "${COMOUTwmo}/iscs_0p25_admin_msg"
     if [[ "${SENDDBN_NTC}" == "YES" ]]; then
         "${DBNROOT}/bin/dbn_alert" NTC_LOW WAFS "${job}" "${COMOUTwmo}/wifs_0p25_admin_msg"
         "${DBNROOT}/bin/dbn_alert" NTC_LOW WAFS "${job}" "${COMOUTwmo}/iscs_0p25_admin_msg"
     fi
+fi
 
+if [[ ! -z "$missing_uk_files" ]] ; then
     subject="WARNING! Missing UK data for WAFS GRIB2 0P25 blending, ${PDY} t${cyc}z ${job}"
     echo "*************************************************************" >mailmsg
     echo "*** WARNING! Missing UK data for WAFS GRIB2 0P25 blending ***" >>mailmsg
@@ -54,17 +59,43 @@ if [[ ! -z "$missing_uk_files" ]] ; then
     echo >>mailmsg
     echo "Send alert message to AWC ...... " >>mailmsg
     echo >>mailmsg
-    echo "All missing files:" >>mailmsg
-    echo "------------------" >>mailmsg
     for file in $missing_uk_files ; do
 	cat $file >>mailmsg
     done
 
-    cat mailmsg >"${COMOUT}/${RUN}.t${cyc}z.wafs_blend_0p25_usonly.emailbody"
-    cat "${COMOUT}/${RUN}.t${cyc}z.f${fhr}.wafs_blend_0p25_usonly.emailbody" | mail.py -s "${subject}" "${MAILTO}" -v
+    cat mailmsg >"${COMOUT}/${RUN}.t${cyc}z.wafs_blend_0p25_ukmissing.emailbody"
+    cat "${COMOUT}/${RUN}.t${cyc}z.wafs_blend_0p25_ukmissing.emailbody" | mail.py -s "${subject}" "${MAILTO}" -v
 fi
 
-for fhr in $fhours ; do
-    echo "=================== log file of fhr=$fhr ==================="
-    cat $DATA/${fhr}.log
-done
+
+if [[ ! -z "$missing_us_files" ]] ; then
+    subject="WARNING! Missing US data for WAFS GRIB2 0P25 blending, ${PDY} t${cyc}z ${job}"
+    echo "*************************************************************" >mailmsg
+    echo "*** WARNING! Missing US data for WAFS GRIB2 0P25 blending ***" >>mailmsg
+    echo "*************************************************************" >>mailmsg
+    echo >>mailmsg
+    echo "Send alert message to AWC ...... " >>mailmsg
+    echo >>mailmsg
+    for file in $missing_us_files ; do
+        cat $file >>mailmsg
+    done
+
+    cat mailmsg >"${COMOUT}/${RUN}.t${cyc}z.wafs_blend_0p25_usmissing.emailbody"
+    cat "${COMOUT}/${RUN}.t${cyc}z.wafs_blend_0p25_usmissing.emailbody" | mail.py -s "${subject}" "${MAILTO}" -v
+fi
+
+if [[ ! -z "$no_blending_files" ]] ; then
+    subject="WARNING! Not blended for WAFS GRIB2 0P25 blending, ${PDY} t${cyc}z ${job}"
+    echo "*************************************************************" >mailmsg
+    echo "*** WARNING! Not blended for WAFS GRIB2 0P25 blending     ***" >>mailmsg
+    echo "*************************************************************" >>mailmsg
+    echo >>mailmsg
+    echo "Send alert message to AWC ...... " >>mailmsg
+    echo >>mailmsg
+    for file in $no_blending_files ; do
+        cat $file >>mailmsg
+    done
+
+    cat mailmsg >"${COMOUT}/${RUN}.t${cyc}z.wafs_blend_0p25_noblending.emailbody"
+    cat "${COMOUT}/${RUN}.t${cyc}z.wafs_blend_0p25_noblending.emailbody" | mail.py -s "${subject}" "${MAILTO}" -v
+fi